Fix: Rename DTO to dto and fix breaking imports causing tests to fail
